[MAT-235] separated out jdispatch.hh
