+++ src/jshim/CMakeLists.txt
@@ -11,7 +11,7 @@ SET(CMAKE_C_FLAGS  "${CMAKE_C_FLAGS} -Wl,--no-undefined")
 SET(CMAKE_CC_FLAGS  "${CMAKE_CC_FLAGS} -Wl,--no-undefined -frepo")
  
 
-add_library(jshim SHARED jshim.cc jterminals.cc jbindings.cc exprfactory.cc dispatch.cc jvmmanager.cc)
+add_library(jshim SHARED jshim.cc jterminals.cc jbindings.cc exprfactory.cc jdispatch.cc jvmmanager.cc)
 add_dependencies(jshim rdag)
 target_link_libraries(jshim rdag)
 set_target_properties(jshim PROPERTIES SOVERSION ${longdog_VERSION_MAJOR} VERSION ${longdog_VERSION})
